+++ lib/common.sh
@@ -320,7 +320,8 @@ cache_copy() {
 }
 
 install_jdk() {
-  local install_dir=${1}
+  local install_dir=${1:?}
+  local cache_dir=${2:?}
 
   let start=$(nowms)
   JVM_COMMON_BUILDPACK=${JVM_COMMON_BUILDPACK:-https://buildpack-registry.s3.amazonaws.com/buildpacks/heroku/jvm.tgz}
@@ -332,6 +333,6 @@ install_jdk() {
   mtime "jvm-common.install.time" "${start}"
 
   let start=$(nowms)
-  install_java_with_overlay ${install_dir}
+  install_java_with_overlay "${install_dir}" "${cache_dir}"
   mtime "jvm.install.time" "${start}"
 }