+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Pass cache to JVM install to cache system.properties file
+
 ## v83
 
 * Ensure bash when running sbt-wrapper